backend/requirements.txt
# Web Framework
flask==3.0.2
flask-cors==4.0.0
flask-socketio==5.3.6
gunicorn==21.2.0

# Database
SQLAlchemy==2.0.27
alembic==1.13.1

# HTTP Client
requests==2.31.0

# Utils
python-dotenv==1.0.1
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Testing
pytest==8.0.1
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2 